execution/reddit_analyzer/pyproject.toml
[project]
name = "reddit-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "feedparser>=6.0.12",
    "google-generativeai",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "python-dotenv",
    "markdown",
]
